backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
pandas>=2.2
openpyxl>=3.1
python-pptx>=1.0
httpx>=0.28
python-multipart>=0.0.20
